Revert "refactor(activities): split activity_service.go into domain-specific files"
This reverts commit 45e7c7bff6d850437961aa78625a8e849ebf2df2.
